+++ packages/frontend/src/components/Icon/index.tsx
@@ -39,6 +39,7 @@ export type IconName =
   | 'question_mark'
   | 'reaction'
   | 'reset-image'
+  | 'sell'
   | 'settings'
   | 'swap_vert'
   | 'swap_hor'