+++ packages/frontend/src/components/message/messageFunctions.ts
@@ -141,7 +141,7 @@ export async function downloadFullMessage(messageId: number) {
   await BackendRemote.rpc.downloadFullMessage(selectedAccountId(), messageId)
 }
 
-export async function openWebxdc(messageId: number) {
+export async function openWebxdc(message: Type.Message) {
   const accountId = selectedAccountId()
-  internalOpenWebxdc(accountId, messageId)
+  internalOpenWebxdc(accountId, message)
 }