+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - new icon for Mac users
 - smooth-scroll to newly arriving messages instead of jumping instantly #4125
 - make backup and key export work in browser #4303
+- add ability to add private tags to accounts
 
 ## Changed
 - enable Telegram-style Ctrl + ArrowUp to reply by default #4333
@@ -18,6 +19,9 @@
 - Update `@deltachat/stdio-rpc-server` and `deltachat/jsonrpc-client` to `1.151.2`
 - description parameter in webxdc.sendUpdate is deprecated now #4359
 - Update `@deltachat/stdio-rpc-server` and `deltachat/jsonrpc-client` to `1.151.3`
+- click on WebxdcInfo Message immediately opens webxdc app #4380
+- href can be passed to webxdc sendStatusUpdate to show event related content
+- dev: move responsibility for updating account list to core through the `AccountsChanged` and `AccountsItemChanged` event
 - add dc version, os name and cpu architecture to fatal errors for better error reports #4384
 
 ## Fixed
@@ -27,7 +31,7 @@
 - fix missing linebreaks in quotes #4360
 - avoid showing wrong menu items for blocked users #4353
 - fix: save message draft every 200ms if message text changed #3733
-- fix mac drag window issues #4300
+- fix mac drag window issues #4300 #4385
 - the main window overflowing small screens, or/and if zoom level is high #4156
 - do not clear the draft if sending failed. #4340
 - "Search in \<chat name\>" divider overflowing for long chat names #4375